requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
arrow==1.3.0
asttokens==3.0.0
attrs==24.2.0
autoflake==2.3.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
codecarbon==2.8.1
comm==0.2.2
cryptography==44.0.0
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
deepsearch-glm==1.0.0
Deprecated==1.2.15
dill==0.3.9
dirtyjson==1.0.8
distro==1.9.0
docling==2.10.0
docling-core==2.9.0
docling-ibm-models==2.0.7
docling-parse==3.0.0
easyocr==1.7.2
et_xmlfile==2.0.0
executing==2.1.0
faiss-cpu==1.9.0.post1
fief-client==0.20.0
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.5
idna==3.10
imageio==2.36.1
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jwcrypto==1.5.6
lazy_loader==0.4
llama-cloud==0.1.6
llama-index==0.12.5
llama-index-agent-openai==0.4.0
llama-index-cli==0.4.0
llama-index-core==0.12.5
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.3.4
llama-index-multi-modal-llms-openai==0.4.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.17
lxml==5.3.0
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.2
nltk==3.9.1
numpy==2.2.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.57.1
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt-toolkit==3.0.36
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyflakes==3.2.0
Pygments==2.18.0
pynvml==12.0.0
pypdf==5.1.0
pypdfium2==4.30.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.0.1
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rerankers==0.6.1
rich==13.9.4
rpds-py==0.22.3
Rtree==1.3.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
semchunk==2.2.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.3.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
triton==3.1.0
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.17.0
XlsxWriter==3.2.0
yarl==1.18.3
yaspin==3.1.0
